refactor(lint): move reporters to separate module (#24757)
